backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
Pillow
psycopg2-binary
django-filter
gunicorn
whitenoise
drf-yasg
django-cleanup
dj-database-url